Add utilities for cross-class methods. Combine download and process into single lambda. Add first draft of serverless meetup presentation.
