+++ .gitignore
@@ -6,7 +6,10 @@ node_modules
 __pycache__
 
 src/old
+
+.env
+samconfig.toml
 invasions/tmp
 invasions/src/old
-samconfig.toml
-.env
+invasions/.env
+invasions/.env.json
