+++ invasions/src/layer/irus/__init__.py
@@ -7,4 +7,4 @@ from .ladderrank import IrusLadderRank
 from .ladder import IrusLadder
 from .process import IrusFiles, IrusProcess
 from .report import IrusReport
-
+from .utilities import update_invasions_for_new_member
