+++ invasions/src/process/requirements.txt
@@ -1,0 +1,1 @@
+aws_lambda_powertools
