+++ invasions/step/process.json
@@ -196,7 +196,8 @@
                   "JitterStrategy": "FULL"
               }
           ],
-          "Next": "UpdateInvasionStats"
+          "Next": "UpdateInvasionStats",
+          "ResultPath": "$.FileProcessed"
       },
       "UpdateInvasionStats": {
           "Type": "Task",
@@ -287,7 +288,8 @@
                   "JitterStrategy": "FULL"
               }
           ],
-          "Next": "UpdateMonthlyStats"
+          "Next": "UpdateMonthlyStats",
+          "ResultPath": "$.PostInvasionReportLink"
       },
       "UpdateMonthlyStats": {
           "Type": "Task",
